config/requirements.txt
# Streamlit Application Requirements
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0

# Machine Learning & NLP
torch>=1.9.0
transformers>=4.20.0
tokenizers>=0.12.0
scikit-learn>=1.0.0

# Facebook API & Web Requests
requests>=2.25.0
facebook-sdk>=3.1.0

# Environment & Configuration
python-dotenv>=0.19.0

# Additional utilities
Pillow>=8.0.0
plotly>=5.0.0
